fix: align period to start permision (#13)
* fix: align period to start permision

* fix: monthsElapsed underflow

* refactor: extract _monthPeriodStart, _fixedPeriodStart, and rename _periodDuration

* feat: widen SpendLimit.multiplier from uint8 to uint16

* chore: updated README.md

* test: add unit and integration tests

* test: add more unit and integration test + format code

---------

+++ script/PreparePermission.s.sol
@@ -89,7 +89,7 @@ contract PreparePermission is Script, CodeConstants {
         address token,
         uint160 allowance,
         JustaPermissionManager.PeriodUnit unit,
-        uint8 multiplier
+        uint16 multiplier
     )
         public
         pure
@@ -112,7 +112,7 @@ contract PreparePermission is Script, CodeConstants {
      * @param selector The function selector for the call permission.
      * @param token The token address for the spend limit.
      * @param allowance The spend allowance.
-     * @param periodUnit The period unit (0-6).
+     * @param periodUnit The period unit (0-5).
      * @param multiplier The period multiplier.
      * @return permission The constructed Permission struct.
      */
@@ -127,7 +127,7 @@ contract PreparePermission is Script, CodeConstants {
         address token,
         uint160 allowance,
         uint8 periodUnit,
-        uint8 multiplier
+        uint16 multiplier
     )
         public
         pure
